backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.26.0
requests>=2.31.0
pymongo>=4.6.0
dnspython>=2.4.0

